包管理修正
